requirements.txt
requests
arxiv
openai
arxiv-to-prompt>=0.2.0
PyYAML>=6.0